[Refactor]: displaycontent 하위뷰로 이동
- PlaylistVertical, PlaylistHorizontal 로 직관적으로 구분하기 위해, displaycontent를 Vertical 내로 이동하였습니다.
